Move authorUnion logic to separate file
